pyproject.toml
[tool.poetry]
name = "4 семестр"
version = "0.1.0"
description = ""
authors = ["Bulat Nasyrov <bulatnasirov2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-login = "^0.6.3"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
flask-wtf = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
